pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mri-utils"
dynamic = ["version"]
description = ''
readme = "README.md"
requires-python = ">=3.11"
keywords = []
authors = [{ name = "Patrick Sadil", email = "psadil@gmail.com" }]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "tapipy==1.6.3",
  "ibis-framework[duckdb]==9.3.0",
  "pandas==2.2.2",
]


[tool.hatch.version]
path = "src/mri_utils/__about__.py"

[tool.hatch.envs.default]
installer = "uv"

[tool.hatch.envs.types]
extra-dependencies = ["mypy>=1.0.0"]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/mri_utils tests}"
